Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,26 +1,26 @@
 #include <stdio.h>
 
 int main() {
-    int a,b;
+    double a,b;
     char c;
-    scanf("%d %d",&a,&b);
+    scanf("%lf %lf",&a,&b);
     scanf("%c",&c);
     
     switch(c){
         case '+':
-        printf("%d",a+b);
+        printf("%lf",a+b);
         break
 
         case '-':
-        printf("%d",a-b);
+        printf("%lf",a-b);
         break
 
         case '*':
-        printf("%d",a*b);
+        printf("%lf",a*b);
         break
 
         case '/':
-        printf("%d",a/b);
+        printf("%lf",a/b);
         break
     }
